backend/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.8.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular-sidecar==2025.9.1
gunicorn==23.0.0
idna==3.10
lxml==6.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
